kernel - Try to fix delayed mtime updates on SHARED+RW maps again
* Attempt to fix continuing problems with mtime updates related to
  modifications made via SHARED+RW mmap()s.  As evidenced by builds
  under synth sometimes getting confused.

* Don't update vp->v_lastwrite_ts in vop_stdclose().  Instead, update
  the field and set a new flag VLASTWRITETS upon mmap(SHARED+PROT_WRITE)
  or upon mprotect(PROT_WRITE) on a SHARED mmap.

* Clear the flag on any regular write, utimes, or truncation.

* Adjust various filesystems to update mtime from vp->v_lastwrite_ts
  only upon UIO_NOCOPY writes, and only if VLASTWRITETS is set.

  tmpfs, ufs, hammer, hammer2 adjusted.

+++ sys/vfs/tmpfs/tmpfs.h
@@ -177,12 +177,12 @@ struct tmpfs_node {
 	mode_t			tn_mode;
 	int			tn_flags;
 	nlink_t			tn_links;
-	int32_t			tn_atime;
-	int32_t			tn_atimensec;
-	int32_t			tn_mtime;
-	int32_t			tn_mtimensec;
-	int32_t			tn_ctime;
-	int32_t			tn_ctimensec;
+	long			tn_atime;
+	long			tn_atimensec;
+	long			tn_mtime;
+	long			tn_mtimensec;
+	long			tn_ctime;
+	long			tn_ctimensec;
 	unsigned long		tn_gen;
 	struct lockf		tn_advlock;
